move Box component to separate files
