+++ applications/launchpad_v2/src/components/Box/types.ts
@@ -1,0 +1,13 @@
+import { ReactNode, CSSProperties } from 'react'
+
+type Gradient = {
+  start: string
+  end: string
+}
+
+export type BoxProps = {
+  children: ReactNode
+  border?: boolean
+  style?: CSSProperties
+  gradient?: Gradient
+}
